spark/requirements.txt
google-generativeai==0.7.1
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
confluent-kafka==2.4.0
elasticsearch==8.14.0